#9874: Overload binary assign in unary assign
