+++ docs/source/ttnn/ttnn/ttnn/assign_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.assign_bw:
+
+ttnn.assign_bw
+####################
+
+.. autofunction:: ttnn.assign_bw
