--- docs/source/ttnn/ttnn/ttnn/binary_assign_bw.rst
@@ -1,6 +1,0 @@
-.. _ttnn.binary_assign_bw:
-
-ttnn.binary_assign_bw
-#####################
-
-.. autofunction:: ttnn.binary_assign_bw
