--- docs/source/ttnn/ttnn/ttnn/unary_assign_bw.rst
@@ -1,6 +1,0 @@
-.. _ttnn.unary_assign_bw:
-
-ttnn.unary_assign_bw
-####################
-
-.. autofunction:: ttnn.unary_assign_bw
