+++ tests/tt_eager/python_api_testing/unit_testing/backward_ops/test_backward_assign.py
@@ -17,14 +17,42 @@ from tests.tt_eager.python_api_testing.unit_testing.backward_ops.utility_funcs i
         (torch.Size([1, 3, 320, 384])),
     ),
 )
+def test_bw_unary_assign(input_shapes, device):
+    in_data, input_tensor = data_gen_with_range(input_shapes, -100, 100, device, True)
+    grad_data, grad_tensor = data_gen_with_range(input_shapes, -100, 100, device)
+
+    tt_output_tensor_on_device = ttnn.assign_bw(grad_tensor, input_tensor)
+
+    in_data.retain_grad()
+
+    pyt_y = torch.clone(in_data)
+
+    pyt_y.backward(gradient=grad_data)
+
+    golden_tensor = [in_data.grad]
+
+    status = compare_pcc(tt_output_tensor_on_device, golden_tensor)
+    assert status
+
+
+@pytest.mark.parametrize(
+    "input_shapes",
+    (
+        (torch.Size([1, 1, 32, 32])),
+        (torch.Size([1, 1, 320, 384])),
+        (torch.Size([1, 3, 320, 384])),
+    ),
+)
 def test_bw_binary_assign(input_shapes, device):
     in_data, input_tensor = data_gen_with_range(input_shapes, -100, 100, device, True)
     other_data, other_tensor = data_gen_with_range(input_shapes, -100, 100, device, True)
 
     grad_data, grad_tensor = data_gen_with_range(input_shapes, -100, 100, device)
 
-    tt_output_tensor_on_device = ttnn.binary_assign_bw(grad_tensor, input_tensor, other_tensor)
-
+    tt_output_tensor_on_device = ttnn.assign_bw(grad_tensor, input_tensor, other_tensor)
+    print(tt_output_tensor_on_device)
+    print(grad_tensor)
+    print(input_tensor, other_tensor)
     in_data.retain_grad()
 
     pyt_y = torch.clone(in_data)
