+++ ttnn/cpp/ttnn/operations/eltwise/binary_backward/device/binary_backward_op.cpp
@@ -354,7 +354,7 @@ std::vector<std::optional<Tensor>> _binary_eq_bw_overload(
     return _binary_eq_bw(default_queue_id, grad, input, other, output_mem_config, are_required_outputs, input_grad, other_grad);
 }
 
-std::vector<Tensor> _binary_assign_bw(
+std::vector<Tensor> _assign_bw(
     const Tensor& grad, const Tensor& input, const Tensor& other, const MemoryConfig& output_mem_config) {
     std::vector<Tensor> grad_tensor;
     grad_tensor.emplace_back(grad);
@@ -638,8 +638,8 @@ std::function<std::vector<ttnn::Tensor>(const Tensor&, const Tensor&, const Tens
             return _add_bw_inter;
         case BinaryBackwardOpType::BINARY_EQ_BW:
             return _binary_eq_bw_inter;
-        case BinaryBackwardOpType::BINARY_ASSIGN_BW:
-            return _binary_assign_bw;
+        case BinaryBackwardOpType::ASSIGN_BW:
+            return _assign_bw;
         case BinaryBackwardOpType::BINARY_LE_BW:
             return _binary_le_bw;
         case BinaryBackwardOpType::RSUB_BW:
