+++ ttnn/cpp/ttnn/operations/eltwise/unary_backward/device/unary_backward_op.cpp
@@ -48,7 +48,7 @@ std::vector<Tensor> _clamp_bw(
     return grad_tensor;
 }
 
-std::vector<Tensor> _unary_assign_bw(const Tensor& grad, const Tensor& input, const MemoryConfig& output_mem_config) {
+std::vector<Tensor> _assign_bw(const Tensor& grad, const Tensor& input, const MemoryConfig& output_mem_config) {
     std::vector<Tensor> grad_tensor;
     grad_tensor.emplace_back(grad);
     return grad_tensor;
@@ -76,8 +76,8 @@ std::vector<Tensor> _multigammaln_bw(const Tensor& grad, const Tensor& input, co
 
 std::function<std::vector<ttnn::Tensor>(const Tensor&, const Tensor&, const MemoryConfig&)> get_function_type1(UnaryBackwardOpType OpType){
     switch (OpType) {
-        case UnaryBackwardOpType::UNARY_ASSIGN_BW:
-            return _unary_assign_bw;
+        case UnaryBackwardOpType::ASSIGN_BW:
+            return _assign_bw;
         case UnaryBackwardOpType::MULTIGAMMALN_BW:
             return _multigammaln_bw;
         default:
